changed pmodel from double to integer
git-svn-id: svn+ssh://spartan.csl.sri.com/svn/private/mcsat/trunk@45 64e4521d-a91b-dd11-9388-003048836090

+++ utils.c
@@ -450,7 +450,7 @@ void init_atom_table(atom_table_t *table){
   table->atom = (samp_atom_t **) safe_malloc(table->size * sizeof(samp_atom_t *));
   table->assignment = (samp_truth_value_t *)
     safe_malloc(table->size * sizeof(samp_truth_value_t));
-  table->pmodel = (double *) safe_malloc(table->size * sizeof(int32_t));
+  table->pmodel = (int32_t *) safe_malloc(table->size * sizeof(int32_t));
   uint32_t i;
   for (i = 0; i < table->size; i++)
     table->pmodel[i] = 0;//was -1
